Handle journal invalidate
The journal API will sometimes issue an SD_JOURNAL_INVALIDATE which will
result in a repositioning of the cursor in the journal because message
may have appeared or disappeared. From the man page:

    Programs only interested in a strictly sequential stream
    of log data may treat SD_JOURNAL_INVALIDATE the same way as
    SD_JOURNAL_APPEND, thus ignoring any changes to the log view
    earlier than the old end of the log stream.

If we have configured IgnorePreviousMessages, we want to treat the
invalidate event as an append, as per the man page.

Bugfix: VRVDR-49115