omfwd: omit einval error messages
When a source address is configured for rsyslog on some platforms, the
error message "omfwd: error 22 sending via udp: Invalid argument" may
be seen at startup only. This is even with Ipfreebind=1 allowing bind
to an IP address not yet associated with the interface. This transient
error is due to getifaddrs() failing, amongst others. It is acceptable
simply not to display such messages on a system where the configuration
including that for source address is guaranteed to be correct.

VRVDR-50495